requirements.txt
numpy>=1.26
pandas>=2.2
pyyaml>=6.0
tqdm>=4.66
matplotlib>=3.8
scikit-learn>=1.4

sentence-transformers>=3.0
torch>=2.2

umap-learn>=0.5.6
hdbscan>=0.8.39

japanize-matplotlib>=1.1.3
